requirements.txt
mitmproxy~=10.2.2
pycryptodome~=3.20.0
